Luciferase-h65: Prism kNN/region completeness for sparse indices
Fixes Prism k-NN under-returning over a sparse index (a k=2 query returning 1 when the nearest
entities are several empty cells away). Verified ORTHOGONAL to the S0/S1 diagonal: an all-same-half
sparse k-NN failed identically before this change. Discovered during RDR-009 P6 (Luciferase-fok);
pulled into the epic per the Option B re-weigh (the AbstractSpatialIndex signature change below is
the RDR-009 Option B trigger — Option A retained, the change is proportionate and contained).

Root causes (the ASI k-NN machinery assumed integer coordinate space; Prism uses normalized [0,1)
world coordinates):

1. getCellSizeAtLevel returned int and truncated Prism's fractional cell size (worldSize/2^level < 1)
   to 0, so the k-NN expanding-radius fallback seeded its search radius at 0 and radius*=2 stayed 0.
   FIX: widened the abstract getCellSizeAtLevel int->float. Octree/Tetree/SFCArrayIndex return their
   integer cell size widened to float (exact: cell sizes are powers of two below 2^24); Prism returns
   the true fractional size. All callers use `var` and float-compatible arithmetic; the one
   integer-stride caller (grid pre-creation) builds a float Point3f, so float is correct there too.
   The now-redundant Prism.getCellSizeAtLevelFloat delegates to it (deprecated, kept for source compat).

2. Prism.findNodesIntersectingBounds was a stub returning empty, no-op'ing its callers for Prism
   (the k-NN radius search, findCollisionsInRegion, bounded-spanning insert). FIX: a map-scan over
   both prism families collecting nodes whose world AABB intersects the query bounds (mirrors
   entitiesInRegion / rayIntersectAll; consistent with Prism's O(n) k-NN BFS).

3. Even with (1)+(2), the expanding-radius fallback can terminate before covering a sparse index:
   its initial radius is a fine-level cell size and its "no new entities" early-exit + fixed
   expansion cap assume dense clustering near the query, so the true neighbors can sit beyond the
   last ring. FIX: a final full-domain sweep when still short of k — but GATED behind a new
   knnRequiresFullDomainSweep() hook (default false; Prism overrides true). The integer-SFC indices
   (Octree/Tetree/SFCArrayIndex) keep their exact prior behavior: their dedicated SFC range-pruning
   paths + integer-space expansion already guarantee completeness, and an unconditional sweep there
   was both a redundant LITMAX/BIGMIN cost AND a hang (an unbounded/large search box drives the
   integer-SFC interval walk non-terminating — caught by full-suite timeouts during development).
   The sweep's box spans only the root-cell edge (getCellSizeAtLevel(0)) — which already encloses
   the whole domain for any in-domain query — while per-entity acceptance uses the full maxDistance
   (decoupled from the box radius) so far-corner neighbors up to the domain diagonal are not rejected.
   For Prism findNodesIntersectingBounds is an O(n) scan, so the sweep is cheap and bounded.

Tests: PrismSparseKnnCompletenessTest — same-half sparse k-NN regression; sparse cross-diagonal
k-NN; far-corner (domain-diagonal) completeness; fractional cell size. All fail before this change.
Full lucien module green (2441, CI=true), including ForestSpatialQueriesTest.testLargeKValue
(Float.MAX_VALUE, k>n) and the Octree/Tetree/SFCArrayIndex k-NN suites (behavior unchanged — the
sweep is not run for them, and their integer cell sizes widen to the same float values).

+++ lucien/src/main/java/com/hellblazer/luciferase/lucien/octree/Octree.java
@@ -311,7 +311,7 @@ public class Octree<ID extends EntityID, Content> extends AbstractSpatialIndex<M
     }
 
     @Override
-    protected int getCellSizeAtLevel(byte level) {
+    protected float getCellSizeAtLevel(byte level) {
         return Constants.lengthAtLevel(level);
     }
